// src/interfaces/communityMembers.interface.ts
export interface ICommunityMembersAttributes {
    id?: number;
    displayName?: string;
    activityRegionAbroad?: string;
    averageScore?: string;
    gender?: string;
    email?: string;
    phone?: string;
    furigana?: string;
    activityRegion?: string;
    photo?: string;
    businessCard?: string;
    fullName?: string;
}